Avoid cyclically mutating promise in heartbeat wrapper
...to prevent slowly leaking memory and creating junk on heap.

If the input is already a `Bluebird.Promise`, then
`(bluebird.)Promise.resolve(promise)` is an identity function, so
actions on `wrappedPromise` are the same as on `promise`. Each
`.timeout` and `.catch` invoked on a bluebird promise mutates it
internally (bluebird does this) while also building new Promise objects
which will cause an ever-growing number of internal references as long
as this timeout-loop is running (ie. original promise is pending).

This pre-commit behavior grows the memory, albeit very slowly because
the allocated bluebird objects in each cycle are tiny.
More annoyingly, it makes debugging difficult - after a job that was
running for a few days, we have 10k+ promises chained together in the
node.js memory dump, almost all of which are created by bluebird's
`.timeout` (and are not GC'd because the original promise is still
pending).